Move get_device_info from adb_handler to adb_monitor
