+++ modelconverter/utils/__init__.py
@@ -1,4 +1,4 @@
-from .adb_handler import AdbHandler, get_device_info
+from .adb_handler import AdbHandler
 from .calibration_data import download_calibration_data
 from .docker_utils import (
     check_docker,
@@ -36,7 +36,7 @@ from .nn_archive import (
 from .onnx_tools import ONNXModifier, onnx_attach_normalization_to_inputs
 from .progress_handler import create_progress_handler
 from .subprocess import SubprocessHandle, subprocess_run
-from .adb_monitor import AdbMonitorDSP, AdbMonitorPower
+from .adb_monitor import AdbMonitorDSP, AdbMonitorPower, get_device_info
 
 __all__ = [
     "AdbHandler",
