+++ modelconverter/utils/adb_monitor.py
@@ -5,7 +5,7 @@ import time
 import statistics
 from collections.abc import Sequence
 from loguru import logger
-
+import depthai as dai
 
 @dataclass
 class _BaseAdbMonitor(ABC):
@@ -393,3 +393,27 @@ echo "$dsp_util"
         self.idle_dsp_utilization = self.get_stats() or 0.0
         self.idle_dsp_utilization *= 1.1  # add 10% margin
         logger.info(f"Idle DSP utilization: {self.idle_dsp_utilization:.4f}%")
+
+def mxid_to_adb_id(mxid: str) -> str:
+    if mxid.isdigit():
+        return format(int(mxid), "x")
+    return mxid.encode("ascii").hex()
+
+def get_device_info(device_ip: str | None, device_mxid: str | None) -> tuple[str | None, str | None]:
+    if not device_ip and not device_mxid:
+        return None, None
+
+    if device_mxid:
+        for info in dai.Device.getAllAvailableDevices():
+            if device_mxid == info.getDeviceId():
+                if device_ip and device_ip != info.name:
+                    logger.warning(
+                        f"Both device_mxid and device_ip provided, but they refer to different devices. Using device with device_mxid: {device_mxid} and device_ip: {info.name}."
+                    )
+                return info.name, mxid_to_adb_id(device_mxid)
+    if device_ip:
+        with dai.Device(device_ip) as device:
+            inferred_mxid = device.getDeviceId()
+            return device_ip, mxid_to_adb_id(inferred_mxid)
+
+    return None, None
